Fix compile warnings (and assigning int to NAN)

+++ include/Jets.hh
@@ -342,7 +342,7 @@ class Jets
 {
 public:
   Jets(SmartChain& chain, const std::string& track_prefix = "jet_ga_trk_");
-  Jet getJet(int) const;
+  Jet getJet(size_t) const;
   size_t size() const;
   double eventWeight() const;
 private:
